feat: cache for single launches
